+++ frontend/src/components/ui/button/Button.tsx
@@ -1,11 +1,11 @@
 import { Slot } from "@radix-ui/react-slot";
-import type { VariantProps } from "class-variance-authority";
+import { type VariantProps } from "class-variance-authority";
 import * as React from "react";
 
 import { cn } from "@/utils";
-import { buttonVariants } from "./varviants";
+import { buttonVariants } from "./variants";
 
-interface ButtonProps
+export interface ButtonProps
   extends React.ButtonHTMLAttributes<HTMLButtonElement>,
     VariantProps<typeof buttonVariants> {
   asChild?: boolean;
