+++ frontend/src/components/ui/button/index.tsx
@@ -1,1 +1,1 @@
-export { Button } from "./Button";
+export { Button, type ButtonProps } from "./Button";
