+++ frontend/src/components/ui/pagination.tsx
@@ -6,8 +6,8 @@ import {
 import * as React from "react";
 
 import { ButtonProps } from "@/components/ui/button";
+import { buttonVariants } from "@/components/ui/button/variants";
 import { cn } from "@/utils";
-import { buttonVariants } from "./buttonVariants";
 
 const Pagination = ({ className, ...props }: React.ComponentProps<"nav">) => (
   <nav
